intelliroute_project/requirements.txt
django
djangorestframework
requests
python-dotenv
huggingface_hub
firebase-admin
google-auth
google-auth-oauthlib